refactored the generator functions to their own module
