+++ input/input_file_generator.py
@@ -11,343 +11,18 @@ The module is designed to handle both relativistic and non-relativistic
 calculations and supports flexible configuration of input parameters.
 """
 import argparse
-import os
 from sys import argv
-import sqlite3
-from subprocess import run, CalledProcessError
-from typing import List, Dict, Tuple
+from typing import Tuple
 
+from data.fetch_atomic_info import get_atomic_weights
+from input.generators.kpoints import generate_k_points_section
+from input.generators.sections import *
 from ui.ui_helpers import prompt_input, print_error, print_info, print_success, console, progress_track, print_header
 from utils.file_parser import get_poscar_data
 from ui.print_thanks import print_animated_ascii
 from core.project_setup import initialize_project
 from core.input_handler import get_pseudopotential_files
 from data.models import ProjectSetup
-
-
-class InputGenerationError(Exception):
-    """Custom exception for input file generation errors."""
-    pass
-
-
-class WannierParams:
-    """Store shared parameters between nscf_wannier and wannier input files."""
-
-    def __init__(self):
-        self.normal_nbands = None
-        self.normal_kmesh = None
-        self.soc_nbands = None
-        self.soc_kmesh = None
-
-    def set_params(self, nbands: int, kmesh: tuple[int, int, int], is_soc: bool = False) -> None:
-        """Set parameters for either normal or SOC calculation."""
-        if is_soc:
-            self.soc_nbands = nbands
-            self.soc_kmesh = kmesh
-        else:
-            self.normal_nbands = nbands
-            self.normal_kmesh = kmesh
-
-    def get_params(self, is_soc: bool = False) -> tuple[int | None, tuple[int, int, int] | None]:
-        """Get parameters for either normal or SOC calculation."""
-        if is_soc:
-            return self.soc_nbands, self.soc_kmesh
-        return self.normal_nbands, self.normal_kmesh
-
-
-def get_atomic_weights(element_names: List[str]) -> List[float] | None:
-    """
-    Retrieves the atomic weights for the given element names.
-
-    Args:
-        element_names (list): List of element names.
-
-    Returns:
-        list: List of atomic weights corresponding to the element names.
-    """
-    if not element_names:
-        raise ValueError("Element names list cannot be empty.")
-    try:
-        conn = sqlite3.connect("data/elements.db")
-        cursor = conn.cursor()
-        atomic_weights = []
-
-        for element in element_names:
-            cursor.execute("""
-            SELECT atomic_weight FROM elements WHERE symbol=?;""", (element,))
-            result = cursor.fetchone()
-            atomic_weights.append(result[0] if result else None)
-
-        return atomic_weights
-    except sqlite3.Error as e:
-        raise InputGenerationError(f"Database error: {str(e)}")
-
-    finally:
-        if conn:
-            conn.close()
-
-
-def generate_control_section(
-        calculation_type: str,
-        pseudo_dir: str,
-        project_dir: str,
-        compound_name: str,
-        *,
-        etot_conv_thr: float = 1e-8,
-        forc_conv_thr: float = 1e-6,
-        relativistic: bool = False,
-) -> str:
-    """
-    Generates the &CONTROL section of the input file.
-
-    Args:
-        calculation_type (str): The type of calculation (e.g., 'vc-relax', 'scf').
-        pseudo_dir (str): Path to the pseudopotential directory.
-        project_dir (str): Path to the project directory.
-        compound_name (str): Name of the compound.
-        etot_conv_thr (float, optional): Energy convergence threshold. Defaults to 1e-8.
-        forc_conv_thr (float, optional): Force convergence threshold. Defaults to 1e-6.
-        relativistic (bool, optional): Whether the calculation is relativistic.
-                                     Defaults to False.
-
-    Returns:
-        str: The &CONTROL section of the input file.
-    """
-    pseudo_dir_path = os.path.join(
-        "../../" if relativistic else "../", os.path.relpath(pseudo_dir, project_dir))
-
-    return f"""&CONTROL
-    calculation      = '{calculation_type}'
-    outdir           = './out'
-    pseudo_dir       = '{pseudo_dir_path}'
-    prefix           = '{compound_name}'
-    verbosity        = 'high'
-    etot_conv_thr    = {etot_conv_thr}
-    forc_conv_thr    = {forc_conv_thr}
-    tprnfor          = .true.
-    tstress          = .true.
-/
-"""
-
-
-def generate_system_section(
-        number_of_atoms: int,
-        atom_types: int,
-        *,
-        ecutwfc: int = 50,
-        ecutrho: int = 500,
-        number_of_bands: int = None,
-        relativistic: bool = False,
-) -> str:
-    """
-    Generates the &SYSTEM section of the input file.
-
-    Args:
-        number_of_atoms (int): The number of atoms in the system.
-        atom_types (int): The number of distinct atom types.
-        ecutwfc (int, optional): Plane-wave cutoff energy in Ry. Defaults to 50.
-        ecutrho (int, optional): Charge density cutoff energy in Ry. Defaults to 500.
-        number_of_bands (int, optional): The number of bands. Required for NSCF and Bands calculations.
-                                        Defaults to None.
-        relativistic (bool, optional): Whether the calculation is relativistic.
-                                     Defaults to False.
-
-    Returns:
-        str: The &SYSTEM section of the input file.
-    """
-
-    system_section = f"""&SYSTEM
-    ibrav            = 0
-    nat              = {number_of_atoms}
-    ntyp             = {atom_types}
-    ecutwfc          = {ecutwfc}
-    ecutrho          = {ecutrho}
-"""
-
-    if number_of_bands is not None:
-        system_section += f"    nbnd             = {number_of_bands}\n"
-    if relativistic:
-        system_section += """    lspinorb         = .true.
-    noncolin         = .true.
-/
-"""
-    else:
-        system_section += "/\n"
-
-    return system_section
-
-
-def generate_electrons_section(
-        *,
-        relativistic: bool = False,
-        conv_thr: float = 1e-9,
-        electron_maxstep: int = 500
-) -> str:
-    """
-    Generates the &ELECTRONS section of the input file.
-
-    Args:
-        relativistic (bool, optional): Whether the calculation is relativistic.
-                             Defaults to False.
-        conv_thr (float): The convergence threshold for the electronic self-consistent field.
-        electron_maxstep (int): The maximum number of electronic self-consistent field iterations.
-
-
-    Returns:
-        str: The &ELECTRONS section of the input file.
-    """
-
-    electrons_section = f"""&ELECTRONS
-    conv_thr         = {conv_thr}
-    electron_maxstep = {electron_maxstep}
-"""
-
-    if relativistic:
-        electrons_section += """    mixing_beta      = 0.4
-    startingpot      = 'file'
-/
-"""
-        return electrons_section
-
-    else:
-        electrons_section += "/\n"
-        return electrons_section
-
-
-def generate_atomic_species_section(element_names: List[str],
-                                    pseudo_list: Dict[str, str],
-                                    atomic_weights: List[float]) -> str:
-    """
-    Generates the ATOMIC_SPECIES section of the input file.
-
-    Args:
-        element_names (list): List of element names.
-        pseudo_list (dict): Dictionary of element names with their corresponding pseudopotential files.
-        atomic_weights (list): List of atomic weights.
-
-    Returns:
-        str: The ATOMIC_SPECIES section of the input file.
-    """
-
-    atomic_species_section = "ATOMIC_SPECIES\n"
-    for element, weight, pseudo in zip(element_names, atomic_weights, pseudo_list.values()):
-        atomic_species_section += f"{element:<2}   {weight:>8.4f}    {pseudo}\n"
-
-    return atomic_species_section
-
-
-def generate_atomic_positions_section(atomic_labels: List[str], atomic_positions: List[str]) -> str:
-    """
-    Generates the ATOMIC_POSITIONS section of the input file.
-
-    Args:
-        atomic_labels (list): List of atomic labels.
-        atomic_positions (list): List of atomic positions.
-
-    Returns:
-        str: The ATOMIC_POSITIONS section of the input file.
-    """
-
-    atomic_positions_section = "ATOMIC_POSITIONS crystal\n"
-    for element, position in zip(atomic_labels, atomic_positions):
-        atomic_positions_section += f"{element:<2}    {position}\n"
-
-    return atomic_positions_section
-
-
-def generate_cell_parameters_section(lattice_vectors: List[str]) -> str:
-    """
-    Generates the CELL_PARAMETERS section of the input file.
-
-    Args:
-        lattice_vectors (list): List of lattice vectors.
-
-    Returns:
-        str: The CELL_PARAMETERS section of the input file.
-    """
-
-    cell_parameters_section = "CELL_PARAMETERS angstrom\n"
-    for vector in lattice_vectors:
-        cell_parameters_section += f"    {vector}\n"
-
-    return cell_parameters_section
-
-
-def generate_k_points_section(calculation_type: str,
-                              k_mesh_density: Tuple[int, int, int] = None,
-                              is_wannier=False) -> str | None:
-    """
-    Generates the K_POINTS section of the input files for Quantum ESPRESSO or
-    kpoints section of Wannier90 input file.
-
-    Args:
-        calculation_type (str): The calculation type (e.g., 'scf', 'bands').
-        k_mesh_density (tuple): The K-point mesh density.
-        is_wannier (bool): Whether the calculation is for Wannier case
-
-    Returns:
-        str: The K_POINTS section of the input file.
-
-    Raises:
-        InputGenerationError: If the calculation type is invalid or k_mesh_density is not provided.
-    """
-    valid_calc_types = ("relax", "vc-relax", "scf", "bands", "nscf", "wannier")
-
-    if calculation_type not in valid_calc_types:
-        raise InputGenerationError(f"Invalid calculation type: {calculation_type}. "
-                                   f"Valid types are: {', '.join(valid_calc_types)}")
-
-    try:
-        if len(k_mesh_density) != 3:
-            raise ValueError("k_mesh_density must be three integers separated by spaces.")
-
-    except TypeError:
-        if calculation_type != "bands":
-            raise InputGenerationError(f"K-point mesh density cannot be None for {calculation_type} calculation.")
-
-    if calculation_type in ["relax", "vc-relax", "scf"]:
-        k_points_section = "K_POINTS automatic\n"
-        k_points_section += f" {k_mesh_density[0]} {k_mesh_density[1]} {k_mesh_density[2]} 0 0 0\n"
-
-        return k_points_section
-
-    elif calculation_type == "nscf":
-        if is_wannier:
-            try:
-                k_points_section = run(f"utils/kmesh.pl {k_mesh_density[0]} {k_mesh_density[1]} {k_mesh_density[2]}",
-                                       shell=True, check=True, capture_output=True).stdout.decode("utf-8")
-                return k_points_section
-            except CalledProcessError as e:
-                raise InputGenerationError(
-                    f"An error occurred in running kmesh.pl script:\n {e.stderr.decode('utf-8')}")
-
-        else:
-            k_points_section = "K_POINTS automatic\n"
-            k_points_section += f" {k_mesh_density[0]} {k_mesh_density[1]} {k_mesh_density[2]} 0 0 0\n"
-
-            return k_points_section
-
-    elif calculation_type == "wannier":
-        try:
-            k_points_section = run(f"utils/kmesh.pl {k_mesh_density[0]} {k_mesh_density[1]} {k_mesh_density[2]} wann",
-                                   shell=True, check=True, capture_output=True).stdout.decode("utf-8")
-            return k_points_section
-        except CalledProcessError as e:
-            raise InputGenerationError(f"An error occurred in running kmesh.pl script:\n {e.stderr.decode('utf-8')}")
-
-    elif calculation_type == "bands":
-        k_points_section = """K_POINTS crystal_b
-4
-0.0000000000    0.0000000000    0.0000000000    120 ! Gamma
-0.5000000000    0.0000000000    0.0000000000    120 ! M
-0.3333333333    0.3333333333    0.0000000000    120 ! K
-0.0000000000    0.0000000000    0.0000000000      0 ! Gamma
-"""
-        return k_points_section
-
-    else:
-        raise InputGenerationError(f"Unsupported calculation type: {calculation_type}. "
-                                   f"Valid types are: {', '.join(valid_calc_types)}")
 
 
 def generate_pw_input_file(calculation_type: str,
@@ -828,8 +503,9 @@ if __name__ == "__main__":
     args = parser.parse_args(argv[1:])
     compound_name, poscar_file = args.compound_name, args.poscar_file
 
-    project = initialize_project(compound_name, poscar_file, is_input=True)
+    os.chdir("..")
 
+    project = initialize_project(compound_name, poscar_file=poscar_file, is_input=True)
     pseudo_list, rel_pseudo_list = get_pseudopotential_files(project.compound_data.element_names,
                                                              project.pseudo_dir,
                                                              relativistic=True,
